+++ build.gradle
@@ -28,6 +28,8 @@ repositories {
     mavenCentral()
 }
 
+minecraft.accessTransformers.file rootProject.file('src/main/resources/META-INF/accesstransformer.cfg')
+
 dependencies {
     implementation "net.neoforged:neoforge:${neo_version}"
 }
