+++ src/main/java/de/dafuqs/additionalentityattributes/mixin/common/BlockMixin.java
@@ -8,6 +8,8 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.neoforge.event.level.BlockEvent;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -19,17 +21,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(value = Block.class)
 public abstract class BlockMixin {
 
-	@Unique
-	private Player additionalEntityAttributes$breakingPlayer;
-
-	@ModifyArg(method = "popExperience", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ExperienceOrb;award(Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/phys/Vec3;I)V"))
-	private int additionalEntityAttributes$modifyExperience(int originalXP) {
-		return (int) (originalXP * Support.getExperienceMod(additionalEntityAttributes$breakingPlayer));
-	}
-
-	@Inject(method = "popExperience", at = @At("HEAD"))
-	public void additionalEntityAttributes$saveBreakingPlayer(Level world, Player player, BlockPos pos, BlockState state, @Nullable BlockEntity blockEntity, ItemStack stack, CallbackInfo callbackInfo) {
-		additionalEntityAttributes$breakingPlayer = player;
-	}
-
+//	@Unique
+//	private Player additionalEntityAttributes$breakingPlayer;
+//
+//	@ModifyArg(method = "popExperience", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ExperienceOrb;award(Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/phys/Vec3;I)V"))
+//	private int additionalEntityAttributes$modifyExperience(int originalXP) {
+//		return (int) (originalXP * Support.getExperienceMod(additionalEntityAttributes$breakingPlayer));
+//	}
+//
+//	@Inject(method = "playerDestroy", at = @At("HEAD")) //pop experience is not called here on neo, could cause issues?
+//	public void additionalEntityAttributes$saveBreakingPlayer(Level world, Player player, BlockPos pos, BlockState state, @Nullable BlockEntity blockEntity, ItemStack stack, CallbackInfo callbackInfo) {
+//		additionalEntityAttributes$breakingPlayer = player;
+//	}
 }
